+++ pyiron_continuum/elasticity/eshelby.py
